scrap/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Django==3.2.3
django-jsonfield==1.4.1
dotmap==1.3.30
exceptiongroup==1.0.0rc9
Flask==2.2.3
frozenlist==1.3.1
func-timeout==4.3.5
h11==0.14.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.2
mysqlclient==2.1.1
outcome==1.2.0
pure-python-adb==0.3.0.dev0
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
Werkzeug==2.2.3
wsproto==1.2.0
yapf==0.32.0
yarl==1.8.1
zipp==3.15.0
